Fix preseed late_command problems (ash is not bash)

+++ Virtual/Machine/ISO-maker/Files/For-Target/late_command.sh
@@ -1,16 +1,11 @@
 #!/bin/bash
 
 apt -y update 
-set -x 
-set -v 
-[[ $(which git) ]] && apt -y reinstall git || apt -y install git 
+apt -y reinstall git 
 mkdir -p /usr/local/share/data/GitHub/econ-ark 
-[[ ! -e /usr/local/share/data/GitHub/econ-ark/econ-ark-tools ]] && sudo git clone --depth 1 --branch Make-Installer-ISO-WORKS https://github.com/econ-ark/econ-ark-tools /usr/local/share/data/GitHub/econ-ark/econ-ark-tools/Virtual/Machine/ISO-maker/Files/For-Target 
-cd /usr/local/share/data/GitHub/econ-ark/econ-ark-tools 
-chmod -Rf a+rwx * ./.*[0-z]* 
-if [[ -d /var/local ]]
-then rm -Rf /var/local 
-fi 
+[[ ! -e /usr/local/share/data/GitHub/econ-ark/econ-ark-tools ]] && git clone --depth 1 --branch Make-Installer-ISO-WORKS https://github.com/econ-ark/econ-ark-tools /usr/local/share/data/GitHub/econ-ark/econ-ark-tools 
+chmod -Rf a+rwx /usr/local/share/data/GitHub/econ-ark/econ-ark-tools/* /usr/local/share/data/GitHub/econ-ark/econ-ark-tools/.*[0-z]* 
+[[ -d /var/local ]] && then rm -Rf /var/local 
 if [[ ! -L /var/local ]]
 then rm -Rf /var/local 
 ln -s /usr/local/share/data/GitHub/econ-ark/econ-ark-tools/Virtual/Machine/ISO-maker/Files/For-Target /var/local 
@@ -24,8 +19,8 @@ sd=$(cat /tmp/target-dev)
 rm -f /var/local/Size-To-Make-Is-* 
 touch /var/local/Size-To-Make-Is-$(echo MIN) 
 echo $(echo MIN > /var/local/About_This_Install/machine-size.txt) 
-/bin/bash /var/local/start.sh 
 cp /var/local/Disk/Labels/Econ-ARK.disk_label /Econ-ARK.disk_label 
 cp /var/local/Disk/Labels/Econ-ARK.disk_label_2x /Econ-ARK.disk_label_2x 
 cp /var/local/Disk/Icons/Econ-ARK.VolumeIcon.icns /Econ-ARK.VolumeIcon.icns 
 echo Econ-ARK > /.disk_label.contentDetails
+/bin/bash /var/local/start.sh
